10 min

+++ Group/Assiut University Training - Newcomers/Arrays/A_Summation.cpp
@@ -6,7 +6,7 @@ int main() {
     cin>>n;
     vector<int> vec(n);
     long long sum = 0;
-    for (size_t i = 0; i < n; i++)
+    for (long long i = 0; i < n; i++)
     {
         cin>>vec[i];
         sum += vec[i];
